+++ tests/test_logs.py
@@ -8,36 +8,10 @@ import pytest
 import responses
 from exception_reports.logs import AddS3ExceptionReportFilter, AddExceptionReportFilter, async_exception_handler, DEFAULT_LOGGING_CONFIG, ExceptionReportConfigurationError, \
     AddExceptionDataFilter
-from exception_reports.reporter import ExceptionReporter
 
 
-def test_exception_report_data():
-    class CustomException(Exception):
-        pass
-
-    def a(foo):
-        bar = 'hey there'  # noqa
-        b(foo)
-
-    def b(foo):
-        c(foo)
-
-    def c(foo):
-        green = 93  # noqa
-        raise CustomException('yolo!')
-
-    try:
-        a('hi')
-    except Exception as e:
-        exception_data = ExceptionReporter().get_traceback_data()
-        frames = exception_data['frames']
-
-        assert exception_data['exception_type'] == 'CustomException'
-        assert exception_data['exception_value'] == 'yolo!'
-        assert len(frames) == 4
-        assert exception_data['frames'][-1]['function'] == 'c'
-        local_vars = dict(exception_data['frames'][-1]['vars'])
-        assert local_vars['green'] == '93'
+class SpecialException(Exception):
+    pass
 
 
 def test_s3_filter_requires_setup():
@@ -118,10 +92,6 @@ def test_error_handler_reports_multiple_exceptions(tmpdir):
         a('bar')
     except Exception as e:
         logger.exception("There were multiple problems")
-
-
-class SpecialException(Exception):
-    pass
 
 
 @pytest.mark.xfail(raises=(RuntimeError,))
